+++ lib/td_dd_web/schema/data_structure_link.ex
@@ -6,8 +6,10 @@ defmodule TdDdWeb.Schema.DataStructureLinks do
   use Absinthe.Schema.Notation
 
   import Absinthe.Resolution.Helpers, only: [dataloader: 1]
+  alias TdDdWeb.Resolvers
 
   object :data_structure_link do
+    field :id, non_null(:id)
     field :source_id, non_null(:id)
     field :target_id, non_null(:id)
     field :source, :data_structure, resolve: dataloader(TdDd.DataStructures)
@@ -15,6 +17,7 @@ defmodule TdDdWeb.Schema.DataStructureLinks do
     field :inserted_at, :date
     field :updated_at, :date
     field :labels, list_of(:label)
+    field :_actions, :json, resolve: &Resolvers.DataStructureLinks.actions/3
   end
 
   object :label do