+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 - [TD-5479] External_ref in grants to avoid duplicates in grant table
 - [TD-5560] LINEAGE_NODES_TIMEOUT releases environment variable for
   TdDd.Lineage.GraphData nodes GenServer call
+- [TD-5509]
+  - Data structure link labels graphql query
+  - Data structure link creation using source and target structure IDs
 
 ## [5.0.0] 2023-01-30